moved search component to it's page.
